# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OmniAI API
  description: Privacy-first AI chat backend API - Agent-based architecture
  version: 1.0.0
  contact:
    name: OmniAI Support
    url: https://github.com/storm/OmniAI
servers:
  - url: /api/v1
    description: OmniAI v1 API (recommended)
paths:
...
  /chat:
    post:
      tags: [Chat]
      summary: Send a message (non-streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /chat/stream:
    post:
      tags: [Chat]
      summary: Send a message (streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Streaming response (SSE)
          content:
            text/event-stream:
              schema:
                type: string
  /chat/retry:
    post:
      tags: [Chat]
      summary: Retry a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryRequest'
      responses:
        '200':
          description: Retry initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  status:
                    type: string
  /chat/cancel:
    post:
      tags: [Chat]
      summary: Cancel a running chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: Cancelled successfully

  # Voice Endpoints
...